Fixed more 64-bit warnings.

+++ TSProjectController.m
@@ -214,7 +214,7 @@
 
 #pragma mark -
 
-- (int)outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item
+- (NSInteger)outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item
 {
 	if (item == nil)
 		return [[_project periodDays] count];
